static and ajax servers updated and merged
